Modify method to remove blacklisted from check_links
Cycle through list of links to check for link in blacklisted instead of cycling through list of blacklisted to check for link in links.

This makes it so that if the same link is collected twice but is only blacklisted once, the link will be removed twice. Before this, the link would have only been removed once.